Added table actions for list component
